Centralize has_* to util
