+++ custom_components/ha_carrier/util.py
@@ -6,10 +6,42 @@ from collections.abc import Iterable, Mapping
 import logging
 from typing import Any, overload
 
+from carrier_api import System
 from homeassistant.core import callback
 
 _LOGGER: logging.Logger = logging.getLogger(__name__)
 REDACTED = "**REDACTED**"
+
+HEAT_TYPES: list[str] = [
+    "hp_heat",
+    "electric_heat",
+    "reheat",
+    "loop_pump",
+]
+
+COOL_TYPES: list[str] = [
+    "cooling",
+]
+
+FAN_TYPES: list[str] = [
+    "fan",
+    "fan_gas",
+]
+
+
+def has_heat(carrier_system: System) -> bool:
+    """Return True if the Carrier system supports heat source selection."""
+    return any(getattr(carrier_system.energy, heat_type, False) is True for heat_type in HEAT_TYPES)
+
+
+def has_cool(carrier_system: System) -> bool:
+    """Return True if the Carrier system supports cool source selection."""
+    return any(getattr(carrier_system.energy, cool_type, False) is True for cool_type in COOL_TYPES)
+
+
+def has_fan(carrier_system: System) -> bool:
+    """Return True if the Carrier system supports fan mode selection."""
+    return any(getattr(carrier_system.energy, fan_type, False) is True for fan_type in FAN_TYPES)
 
 
 @overload
